refactoring and testing
